Move field and target reference types in with modules
