+++ src/mirage/middleware/patch-auto-field-resolvers.ts
@@ -1,10 +1,11 @@
 import { GraphQLSchema } from 'graphql';
 import { mirageRootQueryResolver } from '../resolvers/root-query';
 import { mirageObjectResolver } from '../resolvers/object';
-import { ResolverMapMiddleware, TargetReference, PackOptions, ResolverMap } from '../../types';
+import { ResolverMapMiddleware, PackOptions, ResolverMap } from '../../types';
 import { walk } from '../../resolver-map/walk';
 import { isRootQueryType, isRootMutationType, isInternalType } from '../../graphql/utils';
 import { resolverExistsInResolverMap, addResolverToMap } from '../../resolver-map/utils';
+import { TargetReference } from '../../resolver-map/reference/target-reference';
 
 export function patchAutoFieldResolvers(target: TargetReference = ['*', '*']): ResolverMapMiddleware {
   return async (resolverMap: ResolverMap, packOptions: PackOptions): Promise<ResolverMap> => {
